src/backend/embedder/requirements.txt
nats-py==2.7.2
#protobuf==4.25.3
python-dotenv==1.0.1
grpcio==1.63.0
grpcio-tools==1.63.0
#sentence-transformers==2.7.0


# --find-links https://download.pytorch.org/whl/torch_stable.html
# torch==2.3.0
## move to torch+cpu to minize a lot
# torch==2.3.0+cpu installed directly in dockerfile
# #torchvision==2.3.0+cpu
# #torchaudio==2.3.0+cpu
#-f https://download.pytorch.org/whl/torch_stable.html


